Added note about using find_fit to SARS.
The find_fit function in SAGE requires the
specification of a model. With the data given,
the find_fit will not give a nice logistic curve
since there is not a characteristic flat part.
To fix this, we add another parameter to allow
for a horizontal shift. Doing this allows the
find_fit function to determine a nice-looking
logistic function.